fixed a bug
In a previous version ground clause become having a fixed ordering
of its literals (in order to merge the same clauses). In the greedy
strategy of walksat in samplesat, the previous implementation always
chooses the first literal of a clause when there is a tie of dcost.
That causes inaccurate results.

Now we randomly select one when there is a tie.